requirements.txt
python-dotenv==1.0.0
Flask==3.0.0
mysql-connector-python==8.2.0
requests==2.28.1
